photoreport/requirements.txt
configparser==3.5.0
Django==1.10
mysqlclient==1.3.7
djangorestframework==3.4.4
pillow==3.3.0
python-docx==0.8.6